// models/glossary.ts
import { IContentItem, Elements } from '@kentico/kontent-delivery';

/**
 * Generated by '@kentico/kontent-model-generator@4.1.0'
 */
export type Glossary = IContentItem<{
  title: Elements.TextElement;
  description: Elements.CustomElement;
  key_learning_area: Elements.LinkedItemsElement<IContentItem>;
}>;